% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaplot.R
\name{gseaplot1}
\alias{gseaplot1}
\title{Title}
\usage{
gseaplot1(
  res,
  psize = 1,
  lsize = 2.5,
  base_size = 16,
  border = TRUE,
  title = "Test Plot",
  lcol = "red",
  segcol = "black",
  xdis = 10000
)
}
\arguments{
\item{res}{The gsea software results}

\item{psize}{Point size}

\item{lsize}{Line size}

\item{base_size}{Theme base size}

\item{border}{Plot border}

\item{title}{Plot title}

\item{lcol}{Line color}

\item{segcol}{Segment color}

\item{xdis}{x axis distance}
}
\value{

}
\description{
Title
}
