% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otresults.R
\name{otresults}
\alias{otresults}
\title{Title This script can output significant up and down genes and up genes and down genes}
\usage{
otresults(res, log2FoldChange = 1, pvalue = 0.05, name = "test")
}
\arguments{
\item{res}{Your all diff data}

\item{log2FoldChange}{The range you want to filter}

\item{pvalue}{The range you want to filter}

\item{name}{The output file name}
}
\value{

}
\description{
Title This script can output significant up and down genes and up genes and down genes
}
