% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicevolcano.R
\name{volcaplot}
\alias{volcaplot}
\title{This is my first R package to plot volcano figures}
\usage{
volcaplot(
  res,
  log2FoldChange,
  pvalue,
  size,
  xcut,
  ycut,
  col = c("#dd2c00", "#438a5e", "grey"),
  base_size = 16,
  xdis = 10,
  ydis = 45,
  alpha = 0.8,
  border = FALSE,
  gene,
  gen_size = 6,
  gen_stoke_col = "#373a40",
  title = "The title of this plot",
  flat1 = c("#dd2c00", "#438a5e", "grey"),
  flat2 = c("#e7305b", "#96bb7c", "grey"),
  flat3 = c("#d54062", "#519872", "grey"),
  flat4 = c("#be0000", "#161d6f", "grey"),
  flat5 = c("#ff9292", "#aee1e1", "grey"),
  flat6 = c("#be0000", "#f58634", "grey"),
  flat7 = c("#c70039", "#1a508b", "grey"),
  flat8 = c("#bb2205", "#61b15a", "grey")
)
}
\arguments{
\item{res}{The DeSeq2 results}

\item{log2FoldChange}{The log2 Fold Change value you want to choose}

\item{pvalue}{The p value you want to choose}

\item{size}{The point size}

\item{xcut}{The x axis range to plot,offer only one value bigger than zero.}

\item{ycut}{The y axis range to plot}

\item{col}{The point color you want to show,this is color list like c('red','blue','gray') stands for up down and no significant genes,defaults flat1}

\item{base_size}{The plot all text size}

\item{xdis}{The breaks of the x axis line}

\item{ydis}{The breaks of the y axis line}

\item{alpha}{The transparency of the point}

\item{border}{The plot border show (true or false)}

\item{gene}{The gene you want to mark on the figure :is a vector}

\item{gen_size}{The gene name mark size on the plot}

\item{gen_stoke_col}{This the mark gene circular color}

\item{title}{The plot title contents}

\item{flat1}{The default color}

\item{flat2}{The second color}

\item{flat3}{The next color}

\item{flat4}{The next color}

\item{flat5}{The next color}

\item{flat6}{The next color}

\item{flat7}{The next color}

\item{flat8}{The next color}

\item{volcaplot}{The plot function name}
}
\value{

}
\description{
Today,I create my first function,a very useful function.
}
\details{
Title nicevolcano

You can use this function to visualize differential results from DeSeq2 software.
}
\author{
zhang jun
}
